+++ korma-triangulate-pathfind/src/commonMain/kotlin/com/soywiz/korma/triangle/pathfind/SpatialMesh.kt
@@ -48,9 +48,9 @@ class SpatialMesh {
             val tp1 = triangle.p1
             val tp2 = triangle.p2
             val sn = Node(
-                x = ((tp0.x + tp1.x + tp2.x) / 3).toInt().toFloat(),
-                y = ((tp0.y + tp1.y + tp2.y) / 3).toInt().toFloat(),
-                z = 0f,
+                x = ((tp0.x + tp1.x + tp2.x) / 3).toInt().toDouble(),
+                y = ((tp0.y + tp1.y + tp2.y) / 3).toInt().toDouble(),
+                z = 0.0,
                 triangle = triangle,
                 G = 0,
                 H = 0
@@ -71,9 +71,9 @@ class SpatialMesh {
     }
 
     inner class Node(
-        val x: Float,
-        val y: Float,
-        val z: Float,
+        val x: Double,
+        val y: Double,
+        val z: Double,
         val triangle: Triangle,
         var G: Int = 0, // Cost
         var H: Int = 0, // Heuristic
