+++ src/commonMain/kotlin/com/soywiz/korma/geom/Orientation.kt
@@ -4,12 +4,12 @@ enum class Orientation(val value: Int) {
     CW(+1), CCW(-1), COLLINEAR(0);
 
     companion object {
-        private const val EPSILON: Float = 1e-12f
+        private const val EPSILON: Double = 1e-12
 
         fun orient2d(pa: IPoint, pb: IPoint, pc: IPoint): Orientation {
-            val detleft: Float = (pa.x - pc.x) * (pb.y - pc.y)
-            val detright: Float = (pa.y - pc.y) * (pb.x - pc.x)
-            val `val`: Float = detleft - detright
+            val detleft: Double = (pa.x - pc.x) * (pb.y - pc.y)
+            val detright: Double = (pa.y - pc.y) * (pb.x - pc.x)
+            val `val`: Double = detleft - detright
 
             if ((`val` > -EPSILON) && (`val` < EPSILON)) return Orientation.COLLINEAR
             if (`val` > 0) return Orientation.CCW
