+++ src/commonMain/kotlin/com/soywiz/korma/geom/Vector3D.kt
@@ -4,21 +4,21 @@ import com.soywiz.korma.internal.*
 import kotlin.math.*
 
 interface IVector3D {
-    val x: Float
-    val y: Float
-    val z: Float
-    val w: Float
+    val x: Double
+    val y: Double
+    val z: Double
+    val w: Double
 
     companion object {
-        inline operator fun invoke(x: Number, y: Number, z: Number, w: Number): Vector3D = Vector3D(x.toFloat(), y.toFloat(), z.toFloat(), w.toFloat())
+        inline operator fun invoke(x: Number, y: Number, z: Number, w: Number): Vector3D = Vector3D(x.toDouble(), y.toDouble(), z.toDouble(), w.toDouble())
     }
 }
 
-val IVector3D.length: Float get() = sqrt((x * x) + (y * y) + (z * z) + (w * w))
+val IVector3D.length: Double get() = sqrt((x * x) + (y * y) + (z * z) + (w * w))
 
-data class Vector3D(override var x: Float, override var y: Float, override var z: Float, override var w: Float) : IVector3D {
+data class Vector3D(override var x: Double, override var y: Double, override var z: Double, override var w: Double) : IVector3D {
     companion object {
-        inline operator fun invoke(x: Number, y: Number, z: Number, w: Number = 0f): Vector3D = Vector3D(x.toFloat(), y.toFloat(), z.toFloat(), w.toFloat())
+        inline operator fun invoke(x: Number, y: Number, z: Number, w: Number = 0.0): Vector3D = Vector3D(x.toDouble(), y.toDouble(), z.toDouble(), w.toDouble())
     }
     override fun toString(): String = "(${x.niceStr}, ${y.niceStr}, ${z.niceStr}, ${w.niceStr})"
 }
