+++ src/commonMain/kotlin/com/soywiz/korma/internal/Internal.kt
@@ -3,7 +3,7 @@ package com.soywiz.korma.internal
 internal val Float.niceStr: String get() = if (this.toLong().toFloat() == this) "${this.toLong()}" else "$this"
 internal val Double.niceStr: String get() = if (this.toLong().toDouble() == this) "${this.toLong()}" else "$this"
 
-internal infix fun Float.umod(other: Float): Float {
+internal infix fun Double.umod(other: Double): Double {
     val remainder = this % other
     return when {
         remainder < 0 -> remainder + other
